PNG export support linux/macos.
fix #17
- svg2png to svgexport
- in activate, execute 'npm rebuild'
